requirements-dev.txt
appdirs==1.4.3
aspy.yaml==1.3.0
atomicwrites==1.3.0
attrs==19.1.0
bandit==1.6.0
black==19.3b0
certifi==2019.3.9
cfgv==2.0.0
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.3
filelock==3.0.12
gitdb2==2.0.5
GitPython==2.1.11
identify==1.4.3
idna==2.8
importlib-metadata==0.17
importlib-resources==1.0.2
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
nodeenv==1.3.3
pbr==5.2.0
pluggy==0.11.0
pre-commit==1.17.0
py==1.8.0
pytest==4.5.0
pytest-cov==2.7.1
pytest-mock==1.10.4
PyYAML==5.1
requests==2.22.0
six==1.12.0
smmap2==2.0.5
stevedore==1.30.1
toml==0.10.0
tox==3.11.1
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
